requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
redis==5.0.8
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
celery==5.4.0